+++ matchbox/provider.go
@@ -3,12 +3,11 @@ package matchbox
 import (
 	"fmt"
 
-	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
-	"github.com/hashicorp/terraform-plugin-sdk/terraform"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
 
-// Provider returns a ResourceProvider for Matchbox.
-func Provider() terraform.ResourceProvider {
+// Provider returns a Provider for Matchbox.
+func Provider() *schema.Provider {
 	return &schema.Provider{
 		Schema: map[string]*schema.Schema{
 			"endpoint": &schema.Schema{