+++ matchbox/resource_group_test.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/hashicorp/terraform-plugin-sdk/helper/resource"
-	"github.com/hashicorp/terraform-plugin-sdk/terraform"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
 	"github.com/poseidon/matchbox/matchbox/storage/testfakes"
 )
 
@@ -54,13 +54,11 @@ func TestResourceGroup(t *testing.T) {
 		return nil
 	}
 
-	resource.Test(t, resource.TestCase{
-		IsUnitTest: true,
-		Providers:  providers,
+	resource.UnitTest(t, resource.TestCase{
+		Providers: testProviders,
 		Steps: []resource.TestStep{{
 			Config: srv.AddProviderConfig(hcl),
 			Check:  check,
 		}},
 	})
-
 }